Stop creating the repos and add the da-aws-clients
I've changed this to stop creating the repos in terraform as one
mistimed destroy will delete everything from these repos.

I've changed it to use the tdr terraform module which gets the repo as a
data source and adds the secrets in. This lets us manage the secrets
properly without wiping out the repo accidentally.

I've also added the da-aws-clients repo. Technically it's a cross
department repo but it can live here for now.